create login and login_required

+++ todo/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, get_object_or_404
 from django.http import HttpResponseRedirect
 from django.urls import reverse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 from .forms import TodoForm
 from .models import Todo
@@ -17,8 +19,9 @@ def get_showing_todos(request, todos):
     return todos
 
 
+@login_required
 def index(request):
-    todos = Todo.objects.all()
+    todos = Todo.objects.filter(owner=request.user)
     completed_count = todos.filter(is_completed=True).count()
     incomplete_count = todos.filter(is_completed=False).count()
     all_count = todos.count()
@@ -31,6 +34,7 @@ def index(request):
     return render(request, 'index.html', context)
 
 
+@login_required
 def create_todo(request):
     form = TodoForm()
     if request.method == 'POST':
@@ -43,8 +47,13 @@ def create_todo(request):
         todo.title = title
         todo.description = description
         todo.is_completed = True if is_completed == 'on' else False
+        todo.owner = request.user
 
-        todo.save()
+        if todo.owner == request.user:
+            todo.save()
+
+        messages.add_message(request, messages.SUCCESS,
+                             'Todo created successfully')
 
         return HttpResponseRedirect(reverse('todo:todo_detail', kwargs={'id': todo.pk}))
 
@@ -54,6 +63,7 @@ def create_todo(request):
     return render(request, 'create-todo.html', context)
 
 
+@login_required
 def todo_detail(request, id):
     todo = get_object_or_404(Todo, pk=id)
     context = {
@@ -70,17 +80,24 @@ def handel_server_error(request):
     return render(request, 'server-error.html')
 
 
+@login_required
 def todo_delete(request, id):
     todo = get_object_or_404(Todo, pk=id)
     context = {
         'todo': todo,
     }
     if request.method == 'POST':
-        todo.delete()
-        return HttpResponseRedirect(reverse('todo:index'))
+        if todo.owner == request.user:
+            todo.delete()
+            messages.add_message(request, messages.SUCCESS,
+                                 'Todo deleted successfully')
+            return HttpResponseRedirect(reverse('todo:index'))
+        return render(request, 'todo-delete.html', context)
+
     return render(request, 'todo-delete.html', context)
 
 
+@login_required
 def todo_edit(request, id):
     todo = get_object_or_404(Todo, pk=id)
     form = TodoForm(instance=todo)
@@ -98,6 +115,8 @@ def todo_edit(request, id):
         todo.is_completed = True if is_completed == 'on' else False
 
         todo.save()
+
+        messages.add_message(request, messages.SUCCESS, 'Todo update success')
 
         return HttpResponseRedirect(reverse('todo:todo_detail', kwargs={'id': todo.pk}))
     return render(request, 'todo-edit.html', context)
